Embedded point and color directly in vertex.
Using pointers means we need to dynamically allocate the point and color
and then free them, which is the really tricky thing. The only benefit
was that we could easily set triangles up to share vertices, and
vertices to share colors, but we can still do it pretty easily with the
copy functions.